+++ src/freshpy/__init__.py
@@ -4,13 +4,14 @@
 :Synopsis:          This is the ``__init__`` module for the freshpy package
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     29 Jan 2025
+:Modified Date:     02 Jan 2026
 """
 
 from .core import FreshPy
+from .models.enums import TicketSourceType, TicketStatus, TicketPriority
 from .utils import version
 
-__all__ = ['core', 'FreshPy', 'api', 'agents', 'tickets']
+__all__ = ['core', 'FreshPy', 'api', 'agents', 'tickets', 'TicketSourceType', 'TicketStatus', 'TicketPriority']
 
 # Define the package version by pulling from the freshpy.utils.version module
 __version__ = version.get_full_version()