+++ src/freshpy/models/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""
+:Package:           freshpy.models
+:Synopsis:          This is the ``__init__`` module for the freshpy.models modules
+:Created By:        Jeff Shurtliff
+:Last Modified:     Jeff Shurtliff
+:Modified Date:     02 Jan 2026
+"""
+
+from .enums import TicketSourceType, TicketStatus, TicketPriority
+
+__all__ = ['TicketSourceType', 'TicketStatus', 'TicketPriority']